[LO-89] 북마크 상세 조회 서비스 개발

+++ src/main/java/com/meoguri/linkocean/controller/bookmark/dto/GetBookmarkResponse.java
@@ -23,7 +23,7 @@ public final class GetBookmarkResponse {
 	private final boolean isFavorite;
 	private final LocalDateTime updatedAt;
 	private final List<String> tags;
-	private final Map<String, Integer> reactionCount;
+	private final Map<String, Long> reactionCount;
 	private final GetBookmarkProfileResponse profile;
 
 	public static GetBookmarkResponse of(final GetBookmarkResult result) {
